+++ packages/flutter_watchos/lib/flutter_watchos.dart
@@ -25,5 +25,6 @@ export 'src/watchos_ffi_bindings.dart' show WatchOSNativeBindings;
 export 'src/platform_extension.dart'
     show FlutterWatchosPlatform, FlutterWatchosPlatformExt;
 export 'src/haptics.dart' show WatchHaptics, WatchHapticType;
+export 'src/status_bar.dart' show WatchStatusBar;
 export 'src/crown_scroll.dart' show WatchCrownScroll;
 export 'src/crown.dart' show WatchCrown, CrownRotationEvent;